🎨 Move enum API to tools
